+++ argilla/tests/integration/test_export_dataset.py
@@ -22,6 +22,7 @@ from typing import Any, List
 
 import argilla as rg
 import pytest
+from argilla._exceptions import ConflictError, SettingsError
 from huggingface_hub.utils._errors import BadRequestError, FileMetadataError, HfHubHTTPError
 
 _RETRIES = 5
@@ -33,9 +34,10 @@ def dataset(client) -> rg.Dataset:
     settings = rg.Settings(
         fields=[
             rg.TextField(name="text"),
+            rg.ImageField(name="image"),
         ],
         questions=[
-            rg.TextQuestion(name="label", use_markdown=False),
+            rg.LabelQuestion(name="label", labels=["positive", "negative"]),
         ],
     )
     dataset = rg.Dataset(
@@ -53,16 +55,19 @@ def mock_data() -> List[dict[str, Any]]:
     return [
         {
             "text": "Hello World, how are you?",
+            "image": "http://mock.url/image",
             "label": "positive",
             "id": uuid.uuid4(),
         },
         {
             "text": "Hello World, how are you?",
+            "image": "http://mock.url/image",
             "label": "negative",
             "id": uuid.uuid4(),
         },
         {
             "text": "Hello World, how are you?",
+            "image": "http://mock.url/image",
             "label": "positive",
             "id": uuid.uuid4(),
         },
@@ -108,6 +113,7 @@ class TestDiskImportExportMixin:
                 exported_dataset = json.load(f)
 
         assert exported_settings["fields"][0]["name"] == "text"
+        assert exported_settings["fields"][1]["name"] == "image"
         assert exported_settings["questions"][0]["name"] == "label"
 
         assert exported_dataset["name"] == dataset.name
@@ -125,16 +131,20 @@ class TestDiskImportExportMixin:
 
         with TemporaryDirectory() as temp_dir:
             output_dir = dataset.to_disk(path=temp_dir, with_records=with_records_export)
-            new_dataset = rg.Dataset.from_disk(output_dir, client=client, with_records=with_records_import)
+            new_dataset = rg.Dataset.from_disk(
+                output_dir, client=client, with_records=with_records_import, name=f"test_{uuid.uuid4()}"
+            )
 
         if with_records_export and with_records_import:
             for i, record in enumerate(new_dataset.records(with_suggestions=True)):
                 assert record.fields["text"] == mock_data[i]["text"]
+                assert record.fields["image"] == mock_data[i]["image"]
                 assert record.suggestions["label"].value == mock_data[i]["label"]
         else:
             assert len(new_dataset.records.to_list()) == 0
 
         assert new_dataset.settings.fields[0].name == "text"
+        assert new_dataset.settings.fields[1].name == "image"
         assert new_dataset.settings.questions[0].name == "label"
 
 
@@ -175,11 +185,69 @@ class TestHubImportExportMixin:
                 match="Trying to load a dataset `with_records=True` but dataset does not contain any records.",
             ):
                 new_dataset = rg.Dataset.from_hub(
-                    repo_id=repo_id, client=client, with_records=with_records_import, token=token
+                    repo_id=repo_id,
+                    client=client,
+                    with_records=with_records_import,
+                    token=token,
+                    name=f"test_{uuid.uuid4()}",
                 )
         else:
             new_dataset = rg.Dataset.from_hub(
-                repo_id=repo_id, client=client, with_records=with_records_import, token=token
+                repo_id=repo_id,
+                client=client,
+                with_records=with_records_import,
+                token=token,
+                name=f"test_{uuid.uuid4()}",
+            )
+
+        if with_records_import and with_records_export:
+            for i, record in enumerate(new_dataset.records(with_suggestions=True)):
+                assert record.fields["text"] == mock_data[i]["text"]
+                assert record.fields["image"] == mock_data[i]["image"]
+                assert record.suggestions["label"].value == mock_data[i]["label"]
+        else:
+            assert len(new_dataset.records.to_list()) == 0
+
+        assert new_dataset.settings.fields[0].name == "text"
+        assert new_dataset.settings.fields[1].name == "image"
+        assert new_dataset.settings.questions[0].name == "label"
+
+    @pytest.mark.parametrize("with_records_import", [True, False])
+    def test_import_dataset_from_hub_using_settings(
+        self,
+        token: str,
+        dataset: rg.Dataset,
+        client,
+        mock_data: List[dict[str, Any]],
+        with_records_export: bool,
+        with_records_import: bool,
+    ):
+        repo_id = (
+            f"argilla-internal-testing/test_import_dataset_from_hub_using_settings_with_records{with_records_export}"
+        )
+        dataset.records.log(records=mock_data)
+
+        dataset.to_hub(repo_id=repo_id, with_records=with_records_export, token=token)
+        settings = rg.Settings(
+            fields=[
+                rg.TextField(name="text"),
+            ],
+            questions=[
+                rg.LabelQuestion(name="label", labels=["positive", "negative"]),
+                rg.LabelQuestion(name="extra_label", labels=["extra_positive", "extra_negative"]),
+            ],
+        )
+        if with_records_import and not with_records_export:
+            with pytest.warns(
+                expected_warning=UserWarning,
+                match="Trying to load a dataset `with_records=True` but dataset does not contain any records.",
+            ):
+                new_dataset = rg.Dataset.from_hub(
+                    repo_id=repo_id, client=client, with_records=with_records_import, token=token, settings=settings
+                )
+        else:
+            new_dataset = rg.Dataset.from_hub(
+                repo_id=repo_id, client=client, with_records=with_records_import, token=token, settings=settings
             )
 
         if with_records_import and with_records_export:
@@ -191,3 +259,93 @@ class TestHubImportExportMixin:
 
         assert new_dataset.settings.fields[0].name == "text"
         assert new_dataset.settings.questions[0].name == "label"
+
+    @pytest.mark.parametrize("with_records_import", [True, False])
+    def test_import_dataset_from_hub_using_settings(
+        self,
+        token: str,
+        dataset: rg.Dataset,
+        client,
+        mock_data: List[dict[str, Any]],
+        with_records_export: bool,
+        with_records_import: bool,
+    ):
+        repo_id = (
+            f"argilla-internal-testing/test_import_dataset_from_hub_using_settings_with_records{with_records_export}"
+        )
+        mock_dataset_name = f"test_import_dataset_from_hub_using_settings_{uuid.uuid4()}"
+        dataset.records.log(records=mock_data)
+
+        dataset.to_hub(repo_id=repo_id, with_records=with_records_export, token=token)
+        settings = rg.Settings(
+            fields=[
+                rg.TextField(name="text"),
+            ],
+            questions=[
+                rg.LabelQuestion(name="label", labels=["positive", "negative"]),
+                rg.LabelQuestion(name="extra_label", labels=["extra_positive", "extra_negative"]),
+            ],
+        )
+        if with_records_import and not with_records_export:
+            with pytest.warns(
+                expected_warning=UserWarning,
+                match="Trying to load a dataset `with_records=True` but dataset does not contain any records.",
+            ):
+                new_dataset = rg.Dataset.from_hub(
+                    repo_id=repo_id,
+                    client=client,
+                    with_records=with_records_import,
+                    token=token,
+                    settings=settings,
+                    name=mock_dataset_name,
+                )
+        else:
+            new_dataset = rg.Dataset.from_hub(
+                repo_id=repo_id,
+                client=client,
+                with_records=with_records_import,
+                token=token,
+                settings=settings,
+                name=mock_dataset_name,
+            )
+
+        if with_records_import and with_records_export:
+            for i, record in enumerate(new_dataset.records(with_suggestions=True)):
+                assert record.fields["text"] == mock_data[i]["text"]
+                assert record.suggestions["label"].value == mock_data[i]["label"]
+        else:
+            assert len(new_dataset.records.to_list()) == 0
+
+        assert new_dataset.settings.fields[0].name == "text"
+        assert new_dataset.settings.questions[0].name == "label"
+        assert new_dataset.settings.questions[1].name == "extra_label"
+        assert len(new_dataset.settings.questions[1].labels) == 2
+        assert new_dataset.settings.questions[1].labels[0] == "extra_positive"
+        assert new_dataset.settings.questions[1].labels[1] == "extra_negative"
+        assert new_dataset.name == mock_dataset_name
+
+    def test_import_dataset_from_hub_using_wrong_settings(
+        self,
+        token: str,
+        dataset: rg.Dataset,
+        client,
+        mock_data: List[dict[str, Any]],
+        with_records_export: bool,
+    ):
+        repo_id = f"argilla-internal-testing/test_import_dataset_from_hub_using_wrong_settings_with_records_{with_records_export}"
+        dataset.records.log(records=mock_data)
+
+        dataset.to_hub(repo_id=repo_id, with_records=with_records_export, token=token)
+        settings = rg.Settings(
+            fields=[
+                rg.TextField(name="text"),
+            ],
+            questions=[
+                rg.RatingQuestion(name="label", values=[1, 2, 3, 4, 5]),
+            ],
+        )
+        if with_records_export:
+            with pytest.raises(SettingsError):
+                rg.Dataset.from_hub(repo_id=repo_id, client=client, token=token, settings=settings)
+        else:
+            rg.Dataset.from_hub(repo_id=repo_id, client=client, token=token, settings=settings)