+++ argilla/tests/unit/test_record_ingestion.py
@@ -17,6 +17,7 @@ from uuid import uuid4
 import pytest
 
 import argilla as rg
+from argilla._exceptions import RecordsIngestionError
 
 
 @pytest.fixture
@@ -48,6 +49,21 @@ def test_ingest_record_from_dict(dataset):
     record = record_api_models[0]
     assert record.fields["prompt"] == "What is the capital of France?"
     assert record.suggestions[0].value == "positive"
+
+
+def test_ingest_record_from_empty_dict_raises(dataset):
+    with pytest.raises(RecordsIngestionError):
+        dataset.records._ingest_records(
+            records=[
+                {"id": "record_id"},
+            ],
+        )
+    with pytest.raises(RecordsIngestionError):
+        dataset.records._ingest_records(
+            records=[
+                {},
+            ],
+        )
 
 
 def test_ingest_record_from_dict_with_mapped_suggestions(dataset):